feat: agrega api del banco bicentenario
- Se creó una API del Banco Bicentenario para que sea
posible descargar el estado de cuenta
de forma automática.
- Se acomodaron los logs para que sea posible
generar logs usando el usuario ficticio del sistema.

+++ backend/database/migrations/2021_04_06_035145_create_bank_transactions_table.php
@@ -26,8 +26,8 @@ class CreateBankTransactionsTable extends Migration
 				->onUpdate('cascade')
 				->onDelete('cascade');
 			
-			$table->integer('reference');
-			$table->integer('concepto');
+			$table->bigInteger('reference');
+			$table->bigInteger('concepto');
 			$table->decimal('amount', 18, 2);
 			$table->date('date');
 			$table->char('code', 4);
